Adjust encoder layer config
Here we need to remove the FC layer in encoder because the size of the FC layer has changed when the input image shape is [C, H/3, W/3], which is not compatible with the original image shape [C, H, W] during IL training.

I will edit the decoder next so we will have this FC layer back in the decoder.